src/main/python/requirements.txt
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.0
scipy==1.13.1
notebook==7.2.0
matplotlib==3.9.0
seaborn==0.13.2
tqdm==4.65.2
python-dotenv
pyyaml==6.0.1
torch
openai==1.30.5
transformers==4.41.2
flash_attn
